scripts/requirements.txt
# Python dependencies for local Bonk testing

python-dotenv>=0.19.0  # Environment variable loading
